Merge branch 'master' of https://github.com/thresholdsoft/apolloPharmacist into manual_billing

+++ app/src/main/java/com/apollopharmacy/mpospharmacist/ui/additem/adapter/CategoryDiscAdapter.java
@@ -2,13 +2,17 @@ package com.apollopharmacy.mpospharmacist.ui.additem.adapter;
 
 import android.content.Context;
 import android.text.Editable;
+import android.text.InputFilter;
 import android.text.TextWatcher;
 import android.view.LayoutInflater;
+import android.view.View;
 import android.view.ViewGroup;
+import android.view.inputmethod.InputMethodManager;
 
 import com.apollopharmacy.mpospharmacist.R;
 import com.apollopharmacy.mpospharmacist.databinding.CategoryDisplayItemBinding;
 import com.apollopharmacy.mpospharmacist.ui.additem.model.ManualDiscCheckRes;
+import com.apollopharmacy.mpospharmacist.utils.DecimalDigitsInputFilter;
 
 import java.util.ArrayList;
 
@@ -40,6 +44,7 @@ public class CategoryDiscAdapter extends RecyclerView.Adapter<CategoryDiscAdapte
     public void onBindViewHolder(@NonNull CategoryDiscAdapter.ViewHolder holder, int position) {
         ManualDiscCheckRes.DisplayList item = displayListArrayList.get(position);
         holder.displayItemBinding.setDisplayItem(item);
+        holder.displayItemBinding.cashPaymentAmountEdit.setFilters(new InputFilter[] {new DecimalDigitsInputFilter(100,2)});
         holder.displayItemBinding.cashPaymentAmountEdit.addTextChangedListener(new TextWatcher() {
             @Override
             public void beforeTextChanged(CharSequence charSequence, int i, int i1, int i2) {
@@ -48,18 +53,39 @@ public class CategoryDiscAdapter extends RecyclerView.Adapter<CategoryDiscAdapte
 
             @Override
             public void onTextChanged(CharSequence charSequence, int i, int i1, int i2) {
-                if(charSequence.length()>0)
-                displayListArrayList.get(position).setDiscountValue(Integer.valueOf(holder.displayItemBinding.cashPaymentAmountEdit.getText().toString()));
+                if(charSequence.length()>0 && !charSequence.toString().equalsIgnoreCase("."))
+                displayListArrayList.get(position).setDiscountValue(Double.parseDouble(holder.displayItemBinding.cashPaymentAmountEdit.getText().toString()));
             }
 
             @Override
             public void afterTextChanged(Editable editable) {
-                if (editable.toString().length() > 1 && editable.toString().startsWith("0")) {
+                if (editable.toString().length() > 1 && editable.toString().startsWith("0") && editable.toString().startsWith(".")) {
                      editable.delete(0,1);
-                }else if(editable.toString().length() == 0){
-                    editable.append("0");
                 }
+//                else if(editable.toString().length() == 0){
+//                    editable.append("0");
+//                }
                 holder.displayItemBinding.cashPaymentAmountEdit.setSelection(editable.length());
+            }
+        });
+
+        holder.displayItemBinding.cashPaymentAmountEdit.setOnFocusChangeListener(new View.OnFocusChangeListener() {
+            @Override
+            public void onFocusChange(View view, boolean b) {
+                if(holder.displayItemBinding.cashPaymentAmountEdit.hasFocus()){
+                    if(holder.displayItemBinding.cashPaymentAmountEdit.getText().toString().equalsIgnoreCase("0") || holder.displayItemBinding.cashPaymentAmountEdit.getText().toString().equalsIgnoreCase(".0")){
+                        holder.displayItemBinding.cashPaymentAmountEdit.setText("");
+                        holder.displayItemBinding.cashPaymentAmountEdit.setCursorVisible(true);
+                    }
+                    InputMethodManager imm = (InputMethodManager)holder.displayItemBinding.cashPaymentAmountEdit.getContext(). getSystemService(Context.INPUT_METHOD_SERVICE);
+                    if (imm != null) {
+                        imm.showSoftInput(holder.displayItemBinding.cashPaymentAmountEdit, InputMethodManager.SHOW_IMPLICIT);
+                    }
+                }else{
+                    if(holder.displayItemBinding.cashPaymentAmountEdit.getText().toString().isEmpty()){
+                        holder.displayItemBinding.cashPaymentAmountEdit.setText("0");
+                    }
+                }
             }
         });
     }
